refactor: extract parsing module from main.rs
